split router
