+++ src/App.tsx
@@ -2,7 +2,7 @@ import { ChakraProvider } from '@chakra-ui/react';
 import { RouterProvider } from 'react-router-dom';
 import { ConfirmContextProvider } from './component/modal/confirm-modal/ConfirmContext';
 import { SideKickContextProvider } from './component/modules/sidekick/SideKickContext';
-import { router } from './router';
+import { router } from './utils/router';
 
 export function App() {
   return (
