+++ src/page-group/InvoicingPageGroup.tsx
@@ -1,6 +1,4 @@
-import { useEffect } from 'react';
-import { Outlet, useRevalidator } from 'react-router-dom';
-import { relDb } from '../backend';
+import { Outlet } from 'react-router-dom';
 import { InvoicingLayout } from '../component/layout/MainLayout';
 
 const items = [
@@ -14,16 +12,6 @@ const items = [
 ];
 
 export function InvoicingPageGroup() {
-  const revalidator = useRevalidator();
-
-  useEffect(() => {
-    const observer = relDb.changes({ since: 'now', live: true }).on('change', (e) => {
-      console.log('change', e);
-      revalidator.revalidate();
-    });
-    return () => observer.cancel();
-  }, []);
-
   return (
     <InvoicingLayout navItems={items}>
       <Outlet />
