+++ src/utils/router/modules/common.tsx
@@ -1,0 +1,13 @@
+import { Navigate } from 'react-router-dom';
+
+export const visitDefault = (route: string) => {
+  return {
+    index: true,
+    element: (
+      <Navigate
+        to={route}
+        replace
+      />
+    ),
+  };
+};
